projects/payment-gateway/requirements.txt
flask
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.11
Flask-Migrate==2.5.2
click==7.0
alembic==1.3.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Mail==0.9.1
flask-marshmallow==0.11.0
Flask-Script==2.0.6
itsdangerous==1.1.0
marshmallow==2.21.0
marshmallow-sqlalchemy==0.19.0
python-dotenv==0.13.0
requests==2.23.0